Gauss

Description

Gaussian Integral evaluation plays an important role in quantum chemistry. Here functions will be provided to compute the most important integrals involving gaussian-type orbitals (GTOs).

Documentation

factorial :: (Enum a, Num a, Ord a) => a -> a

Calculates the factorial f(x) = x!

factorial2 :: (Eq a, Num a) => a -> a

Calculates the Doublefactorial f(x) = x!!

binom :: Integral a => a -> a -> a

Calculates the binomialcoefficient n over k binom :: Integer -> Integer -> Double

distance :: Vector Double -> Vector Double -> Double

Computes the distance between to vectors: |r_A - r-B|^2

frac :: Double -> Double -> Double

Calculates

center :: Double -> Vector Double -> Double -> Vector Double -> Vector Double

erf :: Double -> Double

errof function calculation

f_0 :: Double -> Double

normCtr :: Ctr -> Double

Calculates normalization factor of contracted Gaussian with arbitrary angular momentum

normPG :: PG -> Double

Calculates normalization factor of primitive Gaussian with arbitrary angular momentum

f :: Int -> Int -> Int -> Double -> Double -> Double

Calculates f_k (l1 l2 pa_x pb_x) used in Gaussian Product Theorem

zipContractionWith :: (PG -> PG -> Double) -> Ctr -> Ctr -> Double

Evaluates a given function for two contractions

constr_matrix :: [Ctr] -> (PG -> PG -> Double) -> Matrix Double

Construct a matrix out of a list of Contractions and a function for two primitive gaussians

s_12 :: PG -> PG -> Double

Calculates overlap integral of two primitive gaussians

t_12 :: PG -> PG -> Double

Calculates kinetic energy integral of two primitive gaussians

overlaps :: Double -> Double -> Vector Double -> Vector Double -> Double

Calculates nuclear attraction integral of two primitive gaussians v_12 :: Double -> Vecotr Double -> PG -> PG -> Double v_12 z c (PG lmn1 alpha1 pos1) (PG lmn2 alpha2 pos2) = z*pi*pre/g where pref = exp (-alpha1*alpha2 * (distance pos1 pos2) /g) g = alpha1 + alpha2

kinetic :: Double -> Double -> Vector Double -> Vector Double -> Double

nuclear :: Double -> Double -> Vector Double -> Vector Double -> Vector Double -> Double -> Double

twoelectron :: Double -> Double -> Double -> Double -> Vector Double -> Vector Double -> Vector Double -> Vector Double -> Double