HF

Description

Module HF.

Documentation

Module HF.Data provides all important datastructures and corresponding functions as well as dictionary lists.

module HF.Data

In ReadWrite important functions are provided to

Read input (Geometry information, Basis set information) and convert to internal datastructures, provided in Data

Write output (Geometry, structure and basis set information)

module HF.ReadWrite

Gaussian Integral evaluation plays an important role in quantum chemistry. In HF.Gauss, functions will be provided to compute the most important integrals involving gaussian-type orbitals (GTOs).

module HF.Gauss