HF.ReadWrite

Description

In ReadWrite important functions are provided to

Read input (Geometry information, Basis set information) and convert to internal datastructures, provided in Data

Write output (Geometry, structure and basis set information)

Read

Basis set data

Basis set data can be read using the EMSL database. This is currently done by using the EMSL_Basis_Set_Exchange_Local Python API.

Create a Mol from basis set files:

Assuming we have a basis set file "C_STO_3G.dat" (created with generateFile), we can construct a Mol datastructure by

>>> let carbon = Mol "SINGLE-Carbon" [(getAtomData "C_STO_3G.dat", fromList [1,0,0])]

generateFile

Arguments

:: [Char]
basis, eg. "STO-6G""

-> [Char]
elementsymbol, eg. "C"

-> [Char]
output filename

-> IO (Maybe Handle, Maybe Handle, Maybe Handle, ProcessHandle)
output file is saved to EMSL_Basis_Set_Exchange_Local/data/

Generates a file with orbital data from the EMSL_Basis_Set_Exchange_Local Python API.

Therefore the EMSL_Basis_Set_Exchange_Local folder has to be in the same working directory as from which this program is beeing executed.

NOTE: Not for every basis type and element a basis set exists. For a list of possible Atoms for a given basis type, run

 ./EMSL_api.py list_atoms --basis=<basistype>

Example:

>>> generateFile "STO-3G" "C" "C_STO_3G.dat"

getAtomData

Arguments

:: [Char]
filename of the file where the basis set for a certain element is stored

-> Atom
Atom datastructure

Reads basis set data from a file into Atom datastructure.

For creating a basis set file, see generateFile.

NOTE: The basis set file has to exist before executing this function.

Example:

>>> getAtomData "C_STO_3G.dat"

Geometry

To provide geometry information, it might be useful to read geometry information from an external file, which was created by another programm (e.g. Molden, for which it has been tested). This makes it easy to design a geometric structure with preguessed bond lengths and dihedral angles, etc.

Examplary procedure to integrate external geometry into a Mol datastructure:

Suppose, we have a file "CO2.xyz" containing the cartesian geometric information.

Read the geometry information, using getgeom:

>>> let list = getgeom "CO2.xyz"

With constr_set_from_file, construct basis set data for the different elements which will be saved to files:

>>> constr_set_from_file list "STO-6G"

With get_mol_from_files, read the basis set data to construct the complete molecule:

>>> let mol = get_mol_from_files list "STO-6G" "CO2"

getgeom

Arguments

:: [Char]
directory

-> [([Char], Vector Double)]
List with element symbols and corresponding position vectors

Reads a file containing geometry information and outputs a list of element symbols, e.g. "C" with corresponding positions.

Especially nice, when creating the geometry with a different programm, e.g. Molden.

NOTE: The file should have the cartesian XYZ-file format that is used in Molden.

constr_set_from_file :: [([Char], Vector Double)] -> [Char] -> IO ()

For a list of element symbols with corresponding positions ([([Char], Vector Double)]) and a given basis set, e.g. "STO-3G", save basis set data to multiple files.

IMPORTANT: The files are saved in the format <basis>_<element_name>.dat.

To construct a list of element symbols with corresponding positions ([([Char], Vector Double)]) from a file, see getgeom.

get_mol_from_files

Arguments

:: [([Char], Vector Double)]
list of element symbols with corresponding positions

-> [Char]
basis set name, eg. "STO-3G"

-> [Char]
molname, eg. "CO2"

-> Mol

For a list of element symbols with corresponding positions a given basis set and a molecule name, read the existing basis set files (which can be created by using constr_set_from_file) and construct a Mol datastructure. NOTE the difference between this function and constr_set_from_file!

Write

molInfoPrint :: Mol -> String

Show all Important Information about a given Mol datatype.

Output:

Molecule Name

Number of Atoms

For each Atom:

Atomname

Number of Orbitals

contracted set descriptions

geometry

Usage Examples:

>>> let mol = Mol "SINGLE-Carbon" [(getAtomData "C_STO_3G.dat", fromList [1,0,0])]

Normal output:

>>> putStrLn $ molInfoprint mol
Molecule name:   SINGLE-Carbon
Number of Atoms:  1
	Atom:    CARBON
	 	 Number of Orbitals:	3
	 	 contracted sets:	[("S",3),("S",3),("P",3)]
	 	 geometry:	 	[1.0,0.0,0.0]

Save output to file:

>>> writeFile filename $ molInfoPrint mol

molGeomPrint :: Mol -> String

Show atoms with corresponding positions Especially useful for viewing the molecule in a different graphical program, s.a. molden

Usage:

Normal output:

>>> putStrLn $ molGeomprint mol
    1
 C	1.000000	0.000000	0.000000

Save output to file:

>>> writeFile filename $ molGeomPrint mol